pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.3",
    "google-genai>=1.70.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-qdrant>=1.1.0",
    "motor>=3.7.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
]
